moved HOUSE_DIALOGUE_MENU into a separate file
